+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.22.0] - (12.10.2023)
+## Changes
+- added possibility load channels by bouquet direct from device
+- config.schema updated
+- cleanup
+
+
 ## [3.20.0] - (18.02.2023)
 ## Changes
 - added Play/Pause functionality in RC